2020-10-28 ==========
fixed a bug that was causing crashes on Linux.
added x-axis lables to the graphs.
added a tab to graph the spending by category

2020-10-24
==========
added error handling for when the user has not selected an input file. sg.Popup to the rescue!

added feature to load new transactions from a bank download CSV file. It reads the file and puts it into a table named "history_download". Then it copies the records to the TransactionList table. Any records that are already in the TransactionList table are silently ignored. (This is due to a setting when the table was created in the SQLite3 database.

2020-10-23
==========
reformatted the transaction table tab to disply 2 decimal places properly

2020-10-22
==========
moved the transaction table to a tab
expanded the transaction and summary tables to 20 rows

2020-10-21
==========
parameterized some of the options in the rungraph function

added a second graph tab - daily spending

Added a graph to the forecast tab